Project Discontinuation
After ~1.5 weeks of studying the possibility of having swiftUI as the principal development framework, the decision of abandoning this framework has been made. This decision was made on a personal level more than on a team level. The framework is in it's infancy, as so, it doesn't satisfy the needs.